updated userCanDeleteMessage and moved getLogMessageBackgroundColor
